refactor: move userDTOs to new package
