+++ rentplace/src/main/java/kattsyn/dev/rentplace/dtos/users/UserDTO.java
@@ -1,10 +1,11 @@
-package kattsyn.dev.rentplace.dtos;
+package kattsyn.dev.rentplace.dtos.users;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
+import kattsyn.dev.rentplace.dtos.images.ImageDTO;
 import kattsyn.dev.rentplace.enums.Gender;
 import kattsyn.dev.rentplace.enums.Role;
 import lombok.AllArgsConstructor;
